+++ packages/neuron-ui/src/containers/Footer/index.tsx
@@ -1,12 +1,14 @@
-import React, { useCallback, useContext } from 'react'
+import React, { useCallback, useContext, useState, useEffect } from 'react'
 import { createPortal } from 'react-dom'
 import { RouteComponentProps } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import { Stack, getTheme, Text, ProgressIndicator, Icon, TooltipHost } from 'office-ui-fabric-react'
+import { Stack, getTheme, Text, ProgressIndicator, Icon, TooltipHost, TeachingBubble } from 'office-ui-fabric-react'
 
+import { openExternal } from 'services/remote'
+import { guideBubbleTimes } from 'services/localCache'
 import { StateWithDispatch } from 'states/stateProvider/reducer'
-import { ConnectionStatus, FULL_SCREENS, Routes } from 'utils/const'
 import { NeuronWalletContext } from 'states/stateProvider'
+import { ConnectionStatus, FULL_SCREENS, RUN_NODE_GUIDE_URL, Routes } from 'utils/const'
 
 const theme = getTheme()
 const stackStyles = {
@@ -31,6 +33,10 @@ export const SyncStatus = ({
     return <Text variant="xSmall">{t('footer.fail-to-fetch-tip-block-number')}</Text>
   }
 
+  if (BigInt(syncedBlockNumber) < BigInt(0)) {
+    return <Text variant="xSmall">{t('footer.sync-not-start')}</Text>
+  }
+
   const percentage = +syncedBlockNumber / +tipBlockNumber
 
   return +syncedBlockNumber + bufferBlockNumber < +tipBlockNumber ? (
@@ -51,7 +57,7 @@ export const SyncStatus = ({
 
 export const NetworkStatus = ({ name, online }: { name: string; online: boolean }) => {
   return (
-    <Stack horizontal verticalAlign="center" tokens={{ childrenGap: 5 }} styles={stackItemStyles}>
+    <Stack id="network-status" horizontal verticalAlign="center" tokens={{ childrenGap: 5 }} styles={stackItemStyles}>
       <Icon
         iconName={online ? 'Connected' : 'Disconnected'}
         styles={{ root: { display: 'flex', alignItems: 'center' } }}
@@ -71,6 +77,24 @@ const Footer = ({
     settings: { networks = [] },
   } = useContext(NeuronWalletContext)
   const [t] = useTranslation()
+  const [showGuide, setShowGuide] = useState(false)
+
+  useEffect(() => {
+    if (connectionStatus !== ConnectionStatus.Online && guideBubbleTimes.getRemaining()) {
+      setShowGuide(true)
+      guideBubbleTimes.reduce()
+    } else {
+      setShowGuide(false)
+    }
+  }, [connectionStatus, setShowGuide])
+
+  const onDismissGuide = useCallback(() => {
+    setShowGuide(false)
+  }, [setShowGuide])
+
+  const onGuideLinkClick = useCallback(() => {
+    openExternal(RUN_NODE_GUIDE_URL)
+  }, [])
 
   const goToNetworksSetting = useCallback(() => {
     history.push(Routes.SettingsNetworks)
@@ -102,6 +126,27 @@ const Footer = ({
         ) : (
           <Text>{t('settings.setting-tabs.network')}</Text>
         )}
+        {showGuide ? (
+          <TeachingBubble
+            target="#network-status"
+            headline={t('messages.run-ckb-guide')}
+            hasSmallHeadline
+            primaryButtonProps={{
+              children: t('common.open'),
+              onClick: onGuideLinkClick,
+            }}
+            onDismiss={onDismissGuide}
+            styles={{
+              subText: {
+                fontSize: '14px',
+              },
+            }}
+          >
+            <Text as="span" variant="xSmall">
+              {t('messages.view-the-run-node-doc')}
+            </Text>
+          </TeachingBubble>
+        ) : null}
       </Stack>
     </Stack>
   )