+++ packages/neuron-ui/src/utils/calculateGlobalAPC.ts
@@ -1,4 +1,4 @@
-import { getBlockByNumber } from '../services/chain'
+import { getHeaderByNumber } from '../services/chain'
 
 const INITIAL_OFFER = BigInt(33600000000)
 const SECONDARY_OFFER = BigInt(1344000000)
@@ -6,24 +6,19 @@ const DAYS_PER_PERIOD = 365 * 4 * 1
 const MILLI_SECONDS_PER_DAY = 24 * 3600 * 1000
 const PERIOD_LENGTH = DAYS_PER_PERIOD * MILLI_SECONDS_PER_DAY
 
-let cachedGenesisTimestamp: number | undefined
-
-export default async (now: number, initialTimestamp: number | undefined = cachedGenesisTimestamp) => {
+export default async (checkPointTimestamp: number, initialTimestamp?: number | undefined) => {
   let genesisTimestamp = initialTimestamp
   if (genesisTimestamp === undefined) {
-    genesisTimestamp = await getBlockByNumber('0x0')
-      .then(b => {
-        cachedGenesisTimestamp = +b.header.timestamp
-        return cachedGenesisTimestamp
-      })
+    genesisTimestamp = await getHeaderByNumber('0x0')
+      .then(h => +h.timestamp)
       .catch(() => undefined)
   }
-  if (genesisTimestamp === undefined || now <= genesisTimestamp) {
+  if (genesisTimestamp === undefined || checkPointTimestamp <= genesisTimestamp) {
     return 0
   }
 
-  const pastPeriods = BigInt(now - genesisTimestamp) / BigInt(PERIOD_LENGTH)
-  const pastDays = Math.ceil(((now - genesisTimestamp) % PERIOD_LENGTH) / MILLI_SECONDS_PER_DAY)
+  const pastPeriods = BigInt(checkPointTimestamp - genesisTimestamp) / BigInt(PERIOD_LENGTH)
+  const pastDays = Math.ceil(((checkPointTimestamp - genesisTimestamp) % PERIOD_LENGTH) / MILLI_SECONDS_PER_DAY)
 
   const realSecondaryOffer =
     BigInt(4) * SECONDARY_OFFER * pastPeriods +