+++ packages/neuron-wallet/src/startup/sync-block-task/indexer.ts
@@ -5,7 +5,6 @@ import IndexerQueue, { LockHashInfo } from 'services/indexer/queue'
 import { Address } from 'database/address/address-dao'
 
 import initConnection from 'database/chain/ormconfig'
-import ChainInfo from 'models/chain-info'
 
 const { nodeService, addressCreatedSubject, walletCreatedSubject } = remote.require('./startup/sync-block-task/params')
 
@@ -19,7 +18,7 @@ export const loadAddressesAndConvert = async (nodeURL: string): Promise<string[]
 
 // call this after network switched
 let indexerQueue: IndexerQueue | undefined
-export const switchNetwork = async (nodeURL: string, genesisBlockHash: string, chain: string) => {
+export const switchNetwork = async (nodeURL: string, genesisBlockHash: string, _chain: string) => {
   // stop all blocks service
   if (indexerQueue) {
     await indexerQueue.stopAndWait()
@@ -27,7 +26,6 @@ export const switchNetwork = async (nodeURL: string, genesisBlockHash: string, c
 
   // disconnect old connection and connect to new database
   await initConnection(genesisBlockHash)
-  ChainInfo.getInstance().setChain(chain)
   // load lockHashes
   const lockHashes: string[] = await loadAddressesAndConvert(nodeURL)
   const lockHashInfos: LockHashInfo[] = lockHashes.map(lockHash => {